fix: remove mixin
