+++ packages/ui/src/pages/page-stations/page-stations.tsx
@@ -1,18 +1,162 @@
-import { Component, h, State, Prop, Mixin } from '@stencil/core';
+import { Component, h, State, Prop } from '@stencil/core';
 import { Swiper } from 'swiper';
 import { Asset, Station } from '@smartcompanion/data';
-import { AudioPlayerUpdate } from '@smartcompanion/services';
+import { AudioPlayerUpdate, ServiceFacade } from '@smartcompanion/services';
 import { getSortedStations } from '../../utils';
-import { audioPlayerBaseComponentFactory } from '../../utils/audio-player-base-component';
 
 @Component({
   tag: 'sc-page-stations',
   styleUrl: 'page-stations.scss',
 })
-export class PageStations extends Mixin(audioPlayerBaseComponentFactory) {
+export class PageStations {
+
+  /** Audio Player Mixin */
+  @State() playing: boolean = false;
+  @State() position = 0;
+  @State() duration = 0;
+  @State() activeIndex = 0;
+  @State() earpiece = false;
+
+  @Prop() facade: ServiceFacade;
+
+  /**
+   * Initialize earpiece/speaker state from local storage
+   * Should be called in componentWillLoad() lifecycle method
+   */
+  async initEarpiece() {
+    this.earpiece = window.localStorage.getItem('audio-earpiece') == 'yes' ? true : false;
+  }
+
+  /**
+   * Initialize eaudio player with given stations
+   * Should be called in componentDidLoad() lifecycle method
+   */
+  async initAudioPlayer(stations: Array<Station>) {
+    await this.facade.getAudioPlayerService().start(stations);
+
+    if (this.earpiece) {
+      await this.facade.getAudioPlayerService().setEarpiece();
+    } else {
+      await this.facade.getAudioPlayerService().setSpeaker();
+    }
+
+    this.facade.getAudioPlayerService().registerUpdateListener(async (update: AudioPlayerUpdate) => {
+      if (update.state == 'playing') {
+        this.playing = true;
+        this.updatePosition();
+      } else if (update.state == 'paused') {
+        this.playing = false;
+      } else if (update.state == 'skip') {
+        this.activeIndex = update.index;
+        await this.onSkip(update);
+        await this.resetAudioPlayer();
+      } else if (update.state == "completed") {
+        await this.onCompleted(update);
+      } else if (update.state == "collected") {
+        const stationId = this.facade.getAudioPlayerService().getStationId(update.index);
+
+        const station = await this
+          .facade
+          .getStationService()
+          .updateCollectedPercentage(
+            stationId,
+            update.id,
+            update.percentage
+          );
+
+        await this.onCollected(update, station);
+      }
+    });
+
+    await this.facade.getAudioPlayerService().select(this.activeIndex);
+    await this.resetAudioPlayer();
+  }
+
+  /**
+   * Destroy audio player and unregister update listener
+   * Should be called in disconnectedCallback() lifecycle method
+   */
+  async destroyAudioPlayer() {
+    this.playing = false;
+    await this.facade.getAudioPlayerService().stop();
+    this.facade.getAudioPlayerService().unregisterUpdateListener();
+  }
+
+  async resetAudioPlayer() {
+    this.playing = false;
+    this.position = 0;
+    this.duration = await this.facade.getAudioPlayerService().getDuration();
+  }
+
+  async select(audioIndex: number) {
+    await this.facade.getAudioPlayerService().select(audioIndex);
+  }
+
+  async next() {
+    await this.facade.getAudioPlayerService().next();
+  }
+
+  async prev() {
+    await this.facade.getAudioPlayerService().prev();
+  }
+
+  async play() {
+    await this.facade.getAudioPlayerService().play();
+  }
+
+  async pause() {
+    await this.facade.getAudioPlayerService().pause();
+  }
+
+  async playPause() {
+    if (this.playing) {
+      await this.pause();
+    } else {
+      await this.play();
+    }
+  }
+
+  async startPositionChange() {
+    await this.facade.getAudioPlayerService().pause();
+  }
+
+  async changePosition(position: number) {
+    this.position = position;
+    await this.facade.getAudioPlayerService().seek(position);
+    await this.facade.getAudioPlayerService().play();
+  }
+
+  updateActiveIndex(id: string, stations: Array<Station>) {
+    const items = this.facade.getAudioPlayerService().getPlayerItems(stations);
+    this.activeIndex = this.facade.getAudioPlayerService().getIndexByStationId(id, items);
+  }
+
+  async toggleOutput() {
+    if (this.earpiece) {
+      await this.facade.getAudioPlayerService().setSpeaker();
+    } else {
+      await this.facade.getAudioPlayerService().setEarpiece();
+    }
+    this.earpiece = !this.earpiece;
+    window.localStorage.setItem('audio-earpiece', this.earpiece ? 'yes' : 'no');
+  }
+
+  async updatePosition() {
+    this.position = await this.facade.getAudioPlayerService().getPosition();
+    this.duration = await this.facade.getAudioPlayerService().getDuration();
+
+    if (this.duration > 0 && this.position < this.duration) {
+      setTimeout(() => {
+        if (this.playing) {
+          this.updatePosition();
+        }
+      }, 800);
+    }
+  }
+  /** End Audio Player Mixin */
 
   protected playerList: Swiper;
-  protected restartPlaying = false;  
+  protected restartPlaying = false;
 
   @State() stations: Array<Station> = [];
 
@@ -26,15 +170,15 @@ export class PageStations extends Mixin(audioPlayerBaseComponentFactory) {
    * The ID of the initial active station to display, if set to null, the first station will be displayed
    */
   @Prop() stationId: string = null;
-  
-  async componentWillLoad() {    
-    this.facade.getMenuService().enable();  
+
+  async componentWillLoad() {
+    this.facade.getMenuService().enable();
     this.stations = await getSortedStations(this.facade, this.tourId);
     this.updateActiveIndex(this.stationId, this.stations);
     this.initEarpiece();
   }
 
-  async componentDidLoad() {    
+  async componentDidLoad() {
     this.playerList = new Swiper('#player-list', {
       direction: 'vertical',
       slidesPerView: 'auto',
@@ -48,7 +192,7 @@ export class PageStations extends Mixin(audioPlayerBaseComponentFactory) {
   async disconnectedCallback() {
     this.destroyAudioPlayer();
   }
-  
+
   getImageUri(index: number, imageIndex: number = 0) {
     const station = this.stations[index];
     return (station.images[imageIndex] as Asset).internalWebUrl;
@@ -78,7 +222,7 @@ export class PageStations extends Mixin(audioPlayerBaseComponentFactory) {
             <ion-fab-button color="light" size="small" onClick={() => this.openMenu()}>
               <ion-icon color="primary" name="menu-outline"></ion-icon>
             </ion-fab-button>
-          </ion-buttons>          
+          </ion-buttons>
         </ion-toolbar>
       </ion-header>,
       <ion-content id="home" fullscreen={true}>
